// src/app/interfaces/aprobacion-servicios/ordenes-servicio-pendientes-confirmar/response-ordenes-servicio-pendientes-confirmar.interface.ts
export interface IResponseOrdServPendConfirmar {
  TipoCompromiso: string;
  NumeroCompromiso: string;
  LineaDetalle: string;
  FechaPreparacion: string;
  Proveedor: string;
  NomProveedor: string;
  Descripcion: string;
  Cantidad: string;
  CantidadRecibida: string;
  CantidadPendiente: string;
  MonedaDocumento: string;
  MontoAfecto: string;
  MontoImpuesto: string;
  MontoTotal: string;
  PrecioUnitario: string;
  IGVExoneradoFlag: string;
}